refactor: extract shared sync executor test helpers to conftest
Consolidate duplicated make_test_config/make_sync_executor_hassette/make_service
helpers and add TEST_TOKEN constant. Addresses clean-code review findings.
